Update README.md
Correct answer is B because traditional project manager manages the project and is a single person responsible for delivering the project. He manages the team and act as a manager where as Scrum master is a coach and servant leader for the team and do not make decisions for team but let team self organize hence there is a lot of difference between Traditional project manager and scrum master.